ci: run the docker-images no-op at the repo root; split persondraft input.go
The no-op job inherited the workflow's backend/ default working
directory with no checkout to create it. And main grew a second
over-cap file (persondraft/input.go, #956), so the whole-tree length
gates fail every backend-touching PR again: the Person360 folding moves
to fold.go, a mechanical move with no behaviour change.
